feat: deprecating terragrunt-info command
